Reorder songs on vote
- pre-test, pre-rebase commit
- if not working will test and commit again
- Sort of hairy:

On vote, we update the upvote/downvote count and thus net.Then we start
a recursive promise chain that will move the song incrementally through
the orderedSongs list while updating the indices of the songs as we swap.
When we reach our new resting place, we finally update the songs index in
our database and make a call to the spotify api to reorder our track in the
Spotify playlist.

Future work: can't can't increment any higher than the currently playing song